+++ WMS.Application/DTOs/Carts/CartItemResponceDto.cs
@@ -2,9 +2,9 @@ namespace WMS.Application.DTOs.Carts;
 
 public class CartItemResponseDto
 {
-    public long CartItemId { get; set; }
+    public int CartItemId { get; set; }
 
-    public long ProductId { get; set; }
+    public int ProductId { get; set; }
 
     public string ProductName { get; set; } = string.Empty;
