+++ WMS.Application/Interfaces/IShipmentService.cs
@@ -6,10 +6,10 @@ namespace WMS.Application.Interfaces;
 
 public interface IShipmentService
 {
-    Task CreateShipmentAsync(MakeShipmentRequestDto dto,long userId);
+    Task CreateShipmentAsync(MakeShipmentRequestDto dto,int userId);
 
     Task<PagedResult<ShipmentResponseDto>>GetAllShipmentsAsync(CommonFilterDto filterDto);
 
-    Task<ShipmentDetailResponseDto>GetShipmentByIdAsync(long shipmentId);
-    Task UpdateShipmentStatusAsync(long shipmentId,ShipmentStatus status,long userId);
+    Task<ShipmentDetailResponseDto>GetShipmentByIdAsync(int shipmentId);
+    Task UpdateShipmentStatusAsync(int shipmentId,ShipmentStatus status,int userId);
 }
