+++ WMS.Domain/Common/BaseEntity.cs
@@ -2,15 +2,15 @@ namespace WMS.Domain.Common;
 
 public abstract class BaseEntity
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
 
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     public DateTime? UpdatedAt { get; set; }
     public DateTime? DeletedAt { get; set; }
 
-    public long? CreatedBy { get; set; }
-    public long? UpdatedBy { get; set; }
-    public long? DeletedBy { get; set; }
+    public int? CreatedBy { get; set; }
+    public int? UpdatedBy { get; set; }
+    public int? DeletedBy { get; set; }
 
     public bool IsDeleted { get; set; } = false;
 }
